services/llm/pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "vLLM Service for Mini Perplexity"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "vllm>=0.11.0",
]
